+++ packages/core/src/editor/Editor.ts
@@ -20,7 +20,6 @@ import EditorPopupMenu from './EditorPopupMenu';
 import UndoManager from '../view/undoable_changes/UndoManager';
 import EditorKeyHandler from './EditorKeyHandler';
 import EventSource from '../view/event/EventSource';
-import Translations from '../i18n/Translations';
 import Client from '../Client';
 import CompactTreeLayout from '../view/layout/CompactTreeLayout';
 import { EditorToolbar } from './EditorToolbar';
@@ -56,9 +55,9 @@ import type ConnectionHandler from '../view/plugins/ConnectionHandler';
 import { show } from '../util/printUtils';
 import type PanningHandler from '../view/plugins/PanningHandler';
 import { cloneCell } from '../util/cellArrayUtils';
-import { TranslationsConfig } from '../i18n/config';
 import type MaxPopupMenu from '../gui/MaxPopupMenu';
 import { isNullish } from '../internal/utils';
+import { isI18nEnabled, translate } from '../internal/i18n-utils';
 
 /**
  * Extends {@link EventSource} to implement an application wrapper for a graph that
@@ -445,7 +444,7 @@ export class Editor extends EventSource {
    * key does not exist then the value is used as the error message. Default is 'askZoom'.
    * @default 'askZoom'
    */
-  askZoomResource = TranslationsConfig.isEnabled() ? 'askZoom' : '';
+  askZoomResource = isI18nEnabled() ? 'askZoom' : '';
 
   // =====================================================================================
   // Group: Controls and Handlers
@@ -456,14 +455,14 @@ export class Editor extends EventSource {
    * this key does not exist then the value is used as the error message. Default is 'lastSaved'.
    * @default 'lastSaved'.
    */
-  lastSavedResource = TranslationsConfig.isEnabled() ? 'lastSaved' : '';
+  lastSavedResource = isI18nEnabled() ? 'lastSaved' : '';
 
   /**
    * Specifies the resource key for the current file info. If the resource for
    * this key does not exist then the value is used as the error message. Default is 'currentFile'.
    * @default 'currentFile'
    */
-  currentFileResource = TranslationsConfig.isEnabled() ? 'currentFile' : '';
+  currentFileResource = isI18nEnabled() ? 'currentFile' : '';
 
   /**
    * Specifies the resource key for the properties window title. If the
@@ -471,7 +470,7 @@ export class Editor extends EventSource {
    * error message. Default is 'properties'.
    * @default 'properties'
    */
-  propertiesResource = TranslationsConfig.isEnabled() ? 'properties' : '';
+  propertiesResource = isI18nEnabled() ? 'properties' : '';
 
   /**
    * Specifies the resource key for the tasks window title. If the
@@ -479,7 +478,7 @@ export class Editor extends EventSource {
    * error message. Default is 'tasks'.
    * @default 'tasks'
    */
-  tasksResource = TranslationsConfig.isEnabled() ? 'tasks' : '';
+  tasksResource = isI18nEnabled() ? 'tasks' : '';
 
   /**
    * Specifies the resource key for the help window title. If the
@@ -487,7 +486,7 @@ export class Editor extends EventSource {
    * error message. Default is 'help'.
    * @default 'help'
    */
-  helpResource = TranslationsConfig.isEnabled() ? 'help' : '';
+  helpResource = isI18nEnabled() ? 'help' : '';
 
   /**
    * Specifies the resource key for the outline window title. If the
@@ -495,7 +494,7 @@ export class Editor extends EventSource {
    * error message. Default is 'outline'.
    * @default 'outline'
    */
-  outlineResource = TranslationsConfig.isEnabled() ? 'outline' : '';
+  outlineResource = isI18nEnabled() ? 'outline' : '';
 
   /**
    * Reference to the {@link MaxWindow} that contains the outline.
@@ -1230,7 +1229,7 @@ export class Editor extends EventSource {
     this.addAction('zoom', (editor: Editor) => {
       const current = editor.graph.getView().scale * 100;
       const preInput = prompt(
-        Translations.get(editor.askZoomResource) || editor.askZoomResource,
+        translate(editor.askZoomResource) || editor.askZoomResource,
         String(current)
       );
 
@@ -1716,16 +1715,14 @@ export class Editor extends EventSource {
       this.addListener(InternalEvent.SAVE, () => {
         const timestamp = new Date().toLocaleString();
         this.setStatus(
-          `${
-            Translations.get(this.lastSavedResource) || this.lastSavedResource
-          }: ${timestamp}`
+          `${translate(this.lastSavedResource) || this.lastSavedResource}: ${timestamp}`
         );
       });
 
       // Updates the statusbar to display the filename when new files are opened
       this.addListener(InternalEvent.OPEN, () => {
         this.setStatus(
-          `${Translations.get(this.currentFileResource) || this.currentFileResource}: ${
+          `${translate(this.currentFileResource) || this.currentFileResource}: ${
             this.filename
           }`
         );
@@ -2040,7 +2037,7 @@ export class Editor extends EventSource {
         // Displays the contents in a window and stores a reference to the
         // window for later hiding of the window
         this.properties = new MaxWindow(
-          Translations.get(this.propertiesResource) || this.propertiesResource,
+          translate(this.propertiesResource) || this.propertiesResource,
           node,
           x,
           y,
@@ -2220,7 +2217,7 @@ export class Editor extends EventSource {
       div.style.paddingLeft = '20px';
       const w = document.body.clientWidth;
       const wnd = new MaxWindow(
-        Translations.get(this.tasksResource) || this.tasksResource,
+        translate(this.tasksResource) || this.tasksResource,
         div,
         w - 220,
         this.tasksTop,
@@ -2282,14 +2279,14 @@ export class Editor extends EventSource {
   /**
    * Shows the help window. If the help window does not exist
    * then it is created using an iframe pointing to the resource
-   * for the <code>urlHelp</code> key or {@link urlHelp} if the resource
+   * for the `urlHelp` key or {@link urlHelp} if the resource
    * is undefined.
    * @param tasks
    */
   showHelp(tasks: any | null = null): void {
     if (this.help == null) {
       const frame = document.createElement('iframe');
-      frame.setAttribute('src', <string>(Translations.get('urlHelp') || this.urlHelp));
+      frame.setAttribute('src', (translate('urlHelp') || this.urlHelp)!);
       frame.setAttribute('height', '100%');
       frame.setAttribute('width', '100%');
       frame.setAttribute('frameBorder', '0');
@@ -2299,7 +2296,7 @@ export class Editor extends EventSource {
       const h = document.body.clientHeight || document.documentElement.clientHeight;
 
       const wnd = new MaxWindow(
-        Translations.get(this.helpResource) || this.helpResource,
+        translate(this.helpResource) || this.helpResource,
         frame,
         (w - this.helpWidth) / 2,
         (h - this.helpHeight) / 3,
@@ -2353,7 +2350,7 @@ export class Editor extends EventSource {
       div.style.cursor = 'move';
 
       const wnd = new MaxWindow(
-        Translations.get(this.outlineResource) || this.outlineResource,
+        translate(this.outlineResource) || this.outlineResource,
         div,
         600,
         480,