+++ packages/core/src/editor/EditorPopupMenu.ts
@@ -19,10 +19,10 @@ limitations under the License.
 import Cell from '../view/cell/Cell';
 import MaxPopupMenu from '../gui/MaxPopupMenu';
 import { getTextContent } from '../util/domUtils';
-import Translations from '../i18n/Translations';
 import Editor from './Editor';
 import { PopupMenuItem } from '../types';
 import { doEval, isNullish } from '../internal/utils';
+import { translate } from '../internal/i18n-utils';
 
 /**
  * Creates popupmenus for mouse events.
@@ -192,7 +192,7 @@ export class EditorPopupMenu {
 
         if (isNullish(condition) || conditions[condition]) {
           let as = item.getAttribute('as')!;
-          as = Translations.get(as) || as;
+          as = translate(as) || as;
           const funct = doEval(getTextContent(<Text>(<unknown>item)));
           const action = item.getAttribute('action');
           let icon = item.getAttribute('icon');