+++ packages/core/src/serialization/codecs/editor/EditorToolbarCodec.ts
@@ -23,8 +23,8 @@ import { convertPoint } from '../../../util/styleUtils';
 import { getClientX, getClientY } from '../../../util/EventUtils';
 import InternalEvent from '../../../view/event/InternalEvent';
 import { getChildNodes, getTextContent } from '../../../util/domUtils';
-import Translations from '../../../i18n/Translations';
 import { doEval, isElement } from '../../../internal/utils';
+import { translate } from '../../../internal/i18n-utils';
 
 /**
  * Custom codec for configuring {@link EditorToolbar}s.
@@ -144,7 +144,7 @@ export class EditorToolbarCodec extends ObjectCodec {
               into.toolbar.addLine();
             } else if (node.nodeName === 'add') {
               let as = <string>node.getAttribute('as');
-              as = Translations.get(as) || as;
+              as = translate(as) || as;
               const icon = node.getAttribute('icon');
               const pressedIcon = node.getAttribute('pressedIcon');
               const action = node.getAttribute('action');