+++ packages/core/src/view/Graph.ts
@@ -45,7 +45,6 @@ import EdgeHandler from './handler/EdgeHandler';
 import VertexHandler from './handler/VertexHandler';
 import EdgeSegmentHandler from './handler/EdgeSegmentHandler';
 import ElbowEdgeHandler from './handler/ElbowEdgeHandler';
-
 import type {
   EdgeStyleFunction,
   GraphPlugin,
@@ -60,8 +59,8 @@ import { registerDefaultEdgeMarkers } from './geometry/edge/MarkerShape';
 import { registerDefaultStyleElements } from './style/register';
 import { applyGraphMixins } from './mixins/_graph-mixins-apply';
 import { getDefaultPlugins } from './plugins';
-import { TranslationsConfig } from '../i18n/config';
 import { isNullish } from '../internal/utils';
+import { isI18nEnabled } from '../internal/i18n-utils';
 
 /**
  * Extends {@link EventSource} to implement a graph component for the browser. This is the main class of the package.
@@ -384,9 +383,7 @@ class Graph extends EventSource {
    * for this key does not exist then the value is used as the error message.
    * @default 'alreadyConnected'
    */
-  alreadyConnectedResource: string = TranslationsConfig.isEnabled()
-    ? 'alreadyConnected'
-    : '';
+  alreadyConnectedResource: string = isI18nEnabled() ? 'alreadyConnected' : '';
 
   /**
    * Specifies the resource key for the warning message to be displayed when
@@ -394,7 +391,7 @@ class Graph extends EventSource {
    * key does not exist then the value is used as the warning message.
    * @default 'containsValidationErrors'
    */
-  containsValidationErrorsResource: string = TranslationsConfig.isEnabled()
+  containsValidationErrorsResource: string = isI18nEnabled()
     ? 'containsValidationErrors'
     : '';