+++ packages/core/src/view/GraphSelectionModel.ts
@@ -23,7 +23,7 @@ import SelectionChange from './undoable_changes/SelectionChange';
 import UndoableEdit from './undoable_changes/UndoableEdit';
 import EventObject from './event/EventObject';
 import InternalEvent from './event/InternalEvent';
-import { TranslationsConfig } from '../i18n/config';
+import { isI18nEnabled } from '../internal/i18n-utils';
 
 /**
  * Implements the selection model for a graph. Here is a listener that handles
@@ -78,7 +78,7 @@ class GraphSelectionModel extends EventSource {
    * the status message.
    * @default 'done'
    */
-  doneResource = TranslationsConfig.isEnabled() ? 'done' : '';
+  doneResource = isI18nEnabled() ? 'done' : '';
 
   /**
    * Specifies the resource key for the status message while the selection is
@@ -86,7 +86,7 @@ class GraphSelectionModel extends EventSource {
    * value is used as the status message.
    * @default 'updatingSelection'
    */
-  updatingSelectionResource = TranslationsConfig.isEnabled() ? 'updatingSelection' : '';
+  updatingSelectionResource = isI18nEnabled() ? 'updatingSelection' : '';
 
   /**
    * Specifies if only one selected item at a time is allowed.